refactor: decouple style updates from react
Removed setProperty, deleteProperty and createBatchUpdates from react
hook. All parameters can be accessed from global stores.
